+++ Sources/Wrappers/AngouriMath.CPP.Importing/CMakeLists.txt
@@ -15,6 +15,6 @@ set(SOURCES
 "AngouriMath.cpp"
 "ErrorCode.cpp")
 
-add_library(${PROJECT_NAME} ${SOURCES} "EntityInstance.h" "FieldCache.h" "EntityInstance.cpp")
+add_library(${PROJECT_NAME} ${SOURCES})
 target_link_directories(${PROJECT_NAME} PUBLIC win-x64)
 target_link_libraries(${PROJECT_NAME} PUBLIC AngouriMath.CPP.Exporting)
