--- Sources/Wrappers/AngouriMath.CPP.Importing/EntityInstance.h
@@ -1,19 +1,0 @@
-#pragma once
-
-#include "TypeAliases.h"
-#include "FieldCache.h"
-#include "AngouriMath.h"
-#include <vector>
-
-namespace AngouriMath::Internal
-{
-    class EntityInstance
-    {
-    private:
-        FieldCache<const std::vector<AngouriMath::Entity>&> nodes;
-    public:
-        EntityRef ref;
-        EntityInstance(EntityRef ref) : ref(ref) { }
-        const std::vector<AngouriMath::Entity>& CachedNodes();
-    };
-}
