+++ Sources/Wrappers/AngouriMath.CPP.Importing/FieldCache.h
@@ -22,5 +22,7 @@ namespace AngouriMath::Internal
             }
             return value;
         }
+
+        FieldCache() { }
     };
 }
